get rid of msct_register_landmarks and move to sct.register.landmarks
